Changed LFS_ERR_CORRUPT to match EILSEQ instead of EBADE
LFS_ERR_CORRUPT is unfortunately not a well defined error code. It's
very important in the context of littlefs, but missing from the standard
error codes defined in Linux.

After some discussions with other developers, it was encouraged to use
the encoding for EILSEQ over EBADE for representing on disk corrupt, as
EILSEQ implies that there is something wrong with the data.

I've changed this now to take advantage of the breaking changes in v2 to
avoid a risky change to a return value.